+++ Examples/Basics/HelloWorld.ps1
@@ -1,0 +1,3 @@
+New-AdaptiveCard {
+    New-CardTextBlock -Text "Hello, World!"
+} | Out-CardPreview